move all data types to types crate
