+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod constants;
-pub mod fetch;
 pub mod init_tunnel;
 mod utils;
 pub mod types;
+pub mod fetch_api;
